# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todo REST API",
        "description": "Простое CRUD API для задач и тегов",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API v1"
        }
    ],
    "paths": {
...
        "/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Список задач",
                "operationId": "8b27a38c561457b4d60e29ab1e62e8b7",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список задач (пагинация Laravel)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseSuccessEnvelope"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Создать задачу",
                "operationId": "66a60b5423eacfb6d12aa49422eac1e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Задача создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskSuccessResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Получить задачу по ID",
                "operationId": "bcc1660e97f3fddfb6f6154cdb30c9bb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Удалить задачу",
                "operationId": "78243eb8e211bc2247da36e1c9ad3b05",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Задача удалена"
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Обновить задачу",
                "operationId": "46265942bf4d16c6f205aad5ed625b31",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Задача обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{taskId}/tags/{tagId}": {
            "post": {
                "tags": [
                    "Tasks",
                    "Tags"
                ],
                "summary": "Прикрепить тег к задаче",
                "operationId": "533c75bb903a52326bac8e46affd03d7",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tagId",
                        "in": "path",
                        "description": "ID тега",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача с обновлёнными тегами",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача или тег не найдены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tasks",
                    "Tags"
                ],
                "summary": "Открепить тег от задачи",
                "operationId": "f32c49ee83b596e8ba34538bb2fb7d4b",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tagId",
                        "in": "path",
                        "description": "ID тега",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Задача с обновлёнными тегами",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача или тег не найдены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tasks/{taskId}/tags": {
            "put": {
                "tags": [
                    "Tasks",
                    "Tags"
                ],
                "summary": "Синхронизировать список тегов задачи",
                "operationId": "5ea91c40acdacb20c48b46675589059b",
                "parameters": [
                    {
                        "name": "taskId",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SyncTagsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Задача с обновлённым списком тегов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskSuccessResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Задача не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Task": {
                "required": [
                    "id",
                    "title",
                    "description",
                    "status",
                    "priority",
                    "due_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Buy milk"
                    },
                    "description": {
                        "type": "string",
                        "example": "Need to buy milk on the way home"
                    },
                    "status": {
                        "description": "Enum TaskStatusEnum",
                        "type": "string",
                        "example": "todo"
                    },
                    "priority": {
                        "description": "Enum TaskPriorityEnum",
                        "type": "string",
                        "example": "high"
                    },
                    "due_at": {
                        "description": "Unix timestamp",
                        "type": "string",
                        "example": "2024-03-09 16:00:00"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Tag"
                        }
                    }
                },
                "type": "object"
            },
            "Tag": {
                "required": [
                    "id",
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "description": {
                        "type": "string",
                        "example": "Personal"
                    }
                },
                "type": "object"
            },
            "StoreTaskRequest": {
                "required": [
                    "title",
                    "description",
                    "status",
                    "priority",
                    "due_at"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "description": "TaskStatusEnum value",
                        "type": "string"
                    },
                    "priority": {
                        "description": "TaskPriorityEnum value",
                        "type": "string"
                    },
                    "due_at": {
                        "description": "Дата/время, парсится в timestamp",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateTaskRequest": {
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "description": "TaskStatusEnum value",
                        "type": "string"
                    },
                    "priority": {
                        "description": "TaskPriorityEnum value",
                        "type": "string"
                    },
                    "due_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SyncTagsRequest": {
                "required": [
                    "tagsIdList"
                ],
                "properties": {
                    "tagsIdList": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "maxItems": 255
                    }
                },
                "type": "object"
            },
...
            "BaseSuccessEnvelope": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {}
                },
                "type": "object"
            },
            "BaseErrorEnvelope": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "data": {}
                },
                "type": "object"
            },
            "TaskSuccessResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseSuccessEnvelope"
                    },
                    {
                        "properties": {
                            "data": {
                                "$ref": "#/components/schemas/Task"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "ErrorResponse": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseErrorEnvelope"
                    },
                    {
                        "properties": {
                            "data": {
                                "type": "string",
                                "example": "Задача с айди 1 не найдена",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "ValidationErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Validation failed. (title field is required)"
                    }
                },
                "type": "object"
            }
        }
    },
...
}